+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - README updated with storybook, changelog, and better description sections + netlify status badge
 - added `Toast` component
+- added `Modal` component
 
 ## [0.0.1] - 2019-05-09